requirements.txt
requests==2.27.1
pytest==6.2.5
allure-pytest==2.9.45
selenium==3.141.0
pytest-xdist==2.5.0